fix(admin): avoid nested transcript backfill steps
